+++ source/Mocka/Classes/ClassMock.php
@@ -4,7 +4,7 @@ namespace Mocka\Classes;
 
 use CodeGenerator\ClassBlock;
 use Mocka\Exception;
-use Mocka\Invokable\Stub;
+use Mocka\Invokables\Invokable\Stub;
 use Mocka\Overrides\Manager;
 use Mocka\Overrides\MethodOverrides\ClassOverrides;
